fix: fixing stuff to work in prod env
