+++ src/pages/elements/create.tsx
@@ -2,14 +2,44 @@
 import { yupResolver } from '@hookform/resolvers/yup'
 import { Button, ElementHeader, Input } from '../../components'
 import { SubmitHandler, useForm } from 'react-hook-form'
-import { schema, FormData } from './initial-state'
+import { FormData } from '../../models/elements-form'
 import Link from 'next/link'
 import useDarkMode from '../../hooks/use-dark-theme'
 import router from 'next/router'
 import { createElement } from '../../services/elements/create'
 import Swal from 'sweetalert2'
+import * as yup from 'yup'
 
 const Create = () => {
+  const schema = yup.object({
+    name: yup
+      .string()
+      .min(3, 'Minimun of 3 characters!')
+      .required('Element name is required!'),
+    number: yup.number().required('Please, provide the atomic number'),
+    atomic_mass: yup.number().required('Atomic Mass is required'),
+    summary: yup.string().nullable(),
+    appearance: yup.string().nullable(),
+    boil: yup.number().optional(),
+    category: yup.string().required('Element category is required'),
+    density: yup.number().optional(),
+    melt: yup.number().optional(),
+    molar_heat: yup.number().optional(),
+    discovered_by: yup.string().nullable(),
+    named_by: yup.string().nullable(),
+    period: yup.number().optional(),
+    xpos: yup.number().optional(),
+    ypos: yup.number().optional(),
+    phase: yup.string().nullable(),
+    symbol: yup.string().nullable(),
+    source: yup.string().nullable(),
+    spectral_img: yup.string().nullable(),
+    electron_configuration: yup.string().nullable(),
+    electron_configuration_semantic: yup.string().nullable(),
+    electron_affinity: yup.number().optional(),
+    element_img: yup.string().nullable(),
+    enabled: yup.bool().default(true)
+  })
   const { isDarkMode } = useDarkMode()
   const { register, handleSubmit, formState, reset, watch } = useForm<FormData>(
     {
